refactor(memory): split persona into package
