+++ tests/unit/test_memory_temporal.py
@@ -375,7 +375,7 @@ def _persona_manager(tmpdir: str):
     from memory.persona import PersonaManager
     cm = _mock_cm(tmpdir)
     with patch("memory.event_log.get_config_manager", return_value=cm), \
-         patch("memory.persona.get_config_manager", return_value=cm):
+         patch("memory.persona.manager.get_config_manager", return_value=cm):
         evl = EventLog()
         evl._config_manager = cm
         pm = PersonaManager(event_log=evl)
@@ -488,7 +488,7 @@ def test_render_past_block_localizes_to_active_language(tmp_path):
     with patch(
         'utils.language_utils.get_global_language', return_value='en',
     ), patch(
-        'memory.persona.get_global_language', return_value='en', create=True,
+        'utils.language_utils.get_global_language', return_value='en',
     ):
         md = pm._compose_markdown_from_trimmed(
             name='Mio',
@@ -564,7 +564,7 @@ async def test_followup_weighted_disabled_uses_list_order(tmp_path):
     cm = _mock_cm(str(tmp_path))
     with patch("memory.event_log.get_config_manager", return_value=cm), \
          patch("memory.facts.get_config_manager", return_value=cm), \
-         patch("memory.persona.get_config_manager", return_value=cm), \
+         patch("memory.persona.manager.get_config_manager", return_value=cm), \
          patch("memory.reflection.manager.get_config_manager", return_value=cm):
         evl = EventLog(); evl._config_manager = cm
         fs = FactStore(); fs._config_manager = cm
@@ -709,7 +709,7 @@ async def test_arecheck_one_legacy_reflection_skips_malformed_head(tmp_path):
     cm = _mock_cm(str(tmp_path))
     with patch("memory.event_log.get_config_manager", return_value=cm), \
          patch("memory.facts.get_config_manager", return_value=cm), \
-         patch("memory.persona.get_config_manager", return_value=cm), \
+         patch("memory.persona.manager.get_config_manager", return_value=cm), \
          patch("memory.reflection.manager.get_config_manager", return_value=cm):
         evl = EventLog(); evl._config_manager = cm
         fs = FactStore(); fs._config_manager = cm
@@ -772,7 +772,7 @@ async def test_arecheck_invalid_scope_bumps_attempts_and_eventually_skips(tmp_pa
     cm = _mock_cm(str(tmp_path))
     with patch("memory.event_log.get_config_manager", return_value=cm), \
          patch("memory.facts.get_config_manager", return_value=cm), \
-         patch("memory.persona.get_config_manager", return_value=cm), \
+         patch("memory.persona.manager.get_config_manager", return_value=cm), \
          patch("memory.reflection.manager.get_config_manager", return_value=cm):
         evl = EventLog(); evl._config_manager = cm
         fs = FactStore(); fs._config_manager = cm
@@ -844,7 +844,7 @@ async def test_followup_weighted_enabled_varies_picks(tmp_path):
     cm = _mock_cm(str(tmp_path))
     with patch("memory.event_log.get_config_manager", return_value=cm), \
          patch("memory.facts.get_config_manager", return_value=cm), \
-         patch("memory.persona.get_config_manager", return_value=cm), \
+         patch("memory.persona.manager.get_config_manager", return_value=cm), \
          patch("memory.reflection.manager.get_config_manager", return_value=cm):
         evl = EventLog(); evl._config_manager = cm
         fs = FactStore(); fs._config_manager = cm
