+++ tests/unit/test_persona_refine_apply.py
@@ -27,7 +27,7 @@ def _install(tmpdir: str):
     from memory.persona import PersonaManager
     cm = _mock_cm(tmpdir)
     with patch("memory.event_log.get_config_manager", return_value=cm), \
-         patch("memory.persona.get_config_manager", return_value=cm):
+         patch("memory.persona.manager.get_config_manager", return_value=cm):
         event_log = EventLog()
         event_log._config_manager = cm
         pm = PersonaManager(event_log=event_log)
