+++ tests/unit/test_persona_version_history.py
@@ -37,7 +37,7 @@ def _install(tmpdir: str):
 
     cm = _mock_cm(tmpdir)
     with patch("memory.event_log.get_config_manager", return_value=cm), \
-         patch("memory.persona.get_config_manager", return_value=cm):
+         patch("memory.persona.manager.get_config_manager", return_value=cm):
         event_log = EventLog()
         event_log._config_manager = cm
         pm = PersonaManager(event_log=event_log)
